Simplify wine override to just the latest staging version
Maintaining all of the patches from GE's fork is a massive PITA, and does not seem to provide too many benefits for the staging version available at the time of this commit (5.21).
If the Proton patches gain a large lead in compatability / performance in the future, then I may go back to integrating the patches from GE's fork.